requirements.txt
# Core dependencies
tensorflow>=2.15.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional for enhanced functionality
wandb>=0.15.0
tensorboard>=2.13.0
jupyter>=1.0.0
ipywidgets>=8.0.0